Fix and Polish and Explain
- Changed my mind about the filename, my IDE was havign a cow, for good reason
- Added efault_level to Callable
- Added Context Manager
- Added examples in documentation
